Flatten the containers up one level, out of their section directories
A directory per Standard section held exactly one entity each, so
xstd/bits/set/finite_bit_set.hpp and xstd/bits/bitset/bitset.hpp spent a path
component saying what the filename already said, and the umbrella above each
re-exported a single header.

    xstd/bits/set/finite_bit_set.hpp -> xstd/bits/finite_bit_set.hpp
    xstd/bits/array/bit_array.hpp    -> xstd/bits/bit_array.hpp
    xstd/bits/bitset/bitset.hpp      -> xstd/bits/bitset.hpp
    xstd/bits/bit/*                  -> xstd/bits/impl/*

set.hpp, array.hpp, bit.hpp and the old bitset.hpp umbrella are gone; bits.hpp
now names the three containers and the views directly. Every container is one
header named after the type it declares, which is what a consumer asks for. The
Standard's sections are still the organizing idea -- they are the rows of the
grid in doc/design.md -- they are just no longer directories.

The vehicles' directory is impl rather than bit because flattening made
bits/bit/array.hpp and bits/bit_array.hpp derive the same include guard, and the
way out of that collision is not the XSTD_SUBDIR_BIT_SUBDIR_ARRAY_HPP spelling
this branch just deleted. The namespace stays xstd::bit: the directory says these
headers are implementation, which is a separate question from what to call the
bit-level operations. No detail namespace -- this library is small enough not to
need one, and its one member would have been the vehicle the grid names anyway.

Verified: 15 of 17 headers compile standalone, the 2 that do not being the same
from_range_t pair as before -- the count fell from 21 and 3 because four
umbrellas went away, one of which was the third failure. clang-tidy over the
regenerated per-header units reports only the two pre-existing
performance-move-const-arg findings. All five view tests still pass, and
test/src/bits/ranges.cpp still builds and runs clean in CI's configuration under
-Weverything -Werror.
