+++ include/xstd/bits.hpp
@@ -6,13 +6,14 @@
 #ifndef XSTD_BITS_HPP
 #define XSTD_BITS_HPP
 
+// Every container is one header named after the type it declares, so a consumer
+// wanting one asks for it by name and this front door is for wanting them all.
+//
 // Not the ext adaptors: each hard-includes the library it adapts, so a front door
 // naming them would put Boost on every consumer path. Ask for those by name.
-
-#include <xstd/bits/array.hpp>  // IWYU pragma: export; bit_array
-#include <xstd/bits/bit.hpp>    // IWYU pragma: export; the block vehicle and the block predicates
-#include <xstd/bits/bitset.hpp> // IWYU pragma: export; bitset
-#include <xstd/bits/ranges.hpp>  // IWYU pragma: export; the views that make a bitset a range
-#include <xstd/bits/set.hpp>    // IWYU pragma: export; finite_bit_set
+#include <xstd/bits/bit_array.hpp>      // IWYU pragma: export; bit_array
+#include <xstd/bits/bitset.hpp>         // IWYU pragma: export; bitset
+#include <xstd/bits/finite_bit_set.hpp> // IWYU pragma: export; finite_bit_set
+#include <xstd/bits/ranges.hpp>         // IWYU pragma: export; set_view, array_view
 
 #endif // XSTD_BITS_HPP
