--- include/xstd/bits/array.hpp
@@ -1,13 +1,0 @@
-//          Copyright Rein Halbersma 2014-2026.
-// Distributed under the Boost Software License, Version 1.0.
-//    (See accompanying file LICENSE_1_0.txt or copy at
-//          http://www.boost.org/LICENSE_1_0.txt)
-
-#ifndef XSTD_BITS_ARRAY_HPP
-#define XSTD_BITS_ARRAY_HPP
-
-// Header <array>, packed: one bit per element rather than one byte.
-
-#include <xstd/bits/array/bit_array.hpp> // IWYU pragma: export; bit_array
-
-#endif // XSTD_BITS_ARRAY_HPP
