--- include/xstd/bits/bit.hpp
@@ -1,16 +1,0 @@
-//          Copyright Rein Halbersma 2014-2026.
-// Distributed under the Boost Software License, Version 1.0.
-//    (See accompanying file LICENSE_1_0.txt or copy at
-//          http://www.boost.org/LICENSE_1_0.txt)
-
-#ifndef XSTD_BITS_BIT_HPP
-#define XSTD_BITS_BIT_HPP
-
-// The storage vehicle every container above is built on, and the block-level
-// operations it is built out of.
-
-#include <xstd/bits/bit/array.hpp>  // IWYU pragma: export; array
-#include <xstd/bits/bit/intrin.hpp> // IWYU pragma: export; countl_zero, countr_zero, popcount
-#include <xstd/bits/bit/pred.hpp>   // IWYU pragma: export; intersects, is_subset_of, not_equal_to
-
-#endif // XSTD_BITS_BIT_HPP
