--- include/xstd/bits/set.hpp
@@ -1,13 +1,0 @@
-//          Copyright Rein Halbersma 2014-2026.
-// Distributed under the Boost Software License, Version 1.0.
-//    (See accompanying file LICENSE_1_0.txt or copy at
-//          http://www.boost.org/LICENSE_1_0.txt)
-
-#ifndef XSTD_BITS_SET_HPP
-#define XSTD_BITS_SET_HPP
-
-// Header <set> over a finite universe of keys, packed one bit per key.
-
-#include <xstd/bits/set/finite_bit_set.hpp> // IWYU pragma: export; finite_bit_set
-
-#endif // XSTD_BITS_SET_HPP
