+++ test/src/array/type_traits.cpp
@@ -3,7 +3,7 @@
 //    (See accompanying file LICENSE_1_0.txt or copy at
 //          http://www.boost.org/LICENSE_1_0.txt)
 
-#include <xstd/bits/array/bit_array.hpp>           // bit_array
+#include <xstd/bits/bit_array.hpp>           // bit_array
 #include <boost/test/unit_test.hpp>     // BOOST_AUTO_TEST_SUITE, BOOST_AUTO_TEST_SUITE_END, BOOST_AUTO_TEST_CASE_TEMPLATE
 #include <array>                        // array
 #include <concepts>                     // equality_comparable, regular, totally_ordered
