+++ test/src/bitset/o1.cpp
@@ -5,7 +5,7 @@
 
 #include <xstd/bits/ext/boost/dynamic_bitset.hpp> // dynamic_bitset
 #include <xstd/bits/ext/std/bitset.hpp>           // bitset
-#include <xstd/bits/bitset/bitset.hpp>              // bitset
+#include <xstd/bits/bitset.hpp>              // bitset
 #include <bitset/exhaustive.hpp>        // all_cardinality_sets, all_singleton_sets, all_valid, any_value, empty_set, full_set
 #include <bitset/primitives.hpp>        // mem_set, mem_reset, mem_bit_not, mem_flip,
                                         // mem_at,
