+++ test/src/bitset/type_traits.cpp
@@ -5,8 +5,8 @@
 
 #include <xstd/bits/ext/boost/dynamic_bitset.hpp> // dynamic_bitset
 #include <xstd/bits/ext/std/bitset.hpp>           // bitset
-#include <xstd/bits/bitset/bitset.hpp>              // bitset
-#include <xstd/bits/set/finite_bit_set.hpp>             // finite_bit_set
+#include <xstd/bits/bitset.hpp>              // bitset
+#include <xstd/bits/finite_bit_set.hpp>             // finite_bit_set
 #include <concepts.hpp>                 // dynamic
 #include <boost/test/unit_test.hpp>     // BOOST_AUTO_TEST_SUITE, BOOST_AUTO_TEST_SUITE_END, BOOST_AUTO_TEST_CASE_TEMPLATE
 #include <concepts>                     // regular, totally_ordered
