+++ test/src/set/o1.cpp
@@ -9,7 +9,7 @@
 #include <set/primitives.hpp>           // constructor, mem_const_reference, mem_const_iterator, mem_front, mem_back,
                                         // mem_empty, mem_size, mem_max_size, mem_insert, mem_erase, mem_clear,
                                         // op_equal, op_compare_three_way, fn_iterator, fn_size, fn_ssize, fn_empty
-#include <xstd/bits/set/finite_bit_set.hpp>             // finite_bit_set
+#include <xstd/bits/finite_bit_set.hpp>             // finite_bit_set
 #include <boost/test/unit_test.hpp>     // BOOST_AUTO_TEST_SUITE, BOOST_AUTO_TEST_SUITE_END, BOOST_AUTO_TEST_CASE_TEMPLATE
 #include <cstddef>                      // size_t
 #include <cstdint>                      // uint8_t, uint16_t, uint32_t, uint64_t
