+++ test/src/set/o4.cpp
@@ -7,7 +7,7 @@
 #include <set/exhaustive.hpp>           // all_doubleton_set_pairs
 #include <set/flat_set.hpp>             // XSTD_TEST_HAS_FLAT_SET
 #include <set/primitives.hpp>           // op_compare_three_way
-#include <xstd/bits/set/finite_bit_set.hpp>             // finite_bit_set
+#include <xstd/bits/finite_bit_set.hpp>             // finite_bit_set
 #include <boost/test/unit_test.hpp>     // BOOST_AUTO_TEST_SUITE, BOOST_AUTO_TEST_SUITE_END, BOOST_AUTO_TEST_CASE_TEMPLATE
 #include <cstddef>                      // size_t
 #include <cstdint>                      // uint8_t, uint16_t, uint32_t, uint64_t
