+++ test/src/set/sieve.cpp
@@ -5,7 +5,7 @@
 
 #include <opt/set/sieve.hpp>            // filter_twins, sift_primes0, sift_primes1
 #include <set/flat_set.hpp>             // XSTD_TEST_HAS_FLAT_SET
-#include <xstd/bits/set/finite_bit_set.hpp>             // finite_bit_set
+#include <xstd/bits/finite_bit_set.hpp>             // finite_bit_set
 #include <boost/test/unit_test.hpp>     // BOOST_AUTO_TEST_SUITE, BOOST_AUTO_TEST_SUITE_END, BOOST_AUTO_TEST_CASE_TEMPLATE
 #include <fmt/format.h>                 // format
 #include <fmt/ranges.h>
